@@ -1,40 +1,39 @@
 ﻿using System;
 using System.Diagnostics.CodeAnalysis;
 
-namespace Pevac
+namespace Pevac;
+
+/// <summary>
+/// Represents a type with a single value.
+/// </summary>
+public readonly struct Void : IEquatable<Void>
 {
     /// <summary>
-    /// Represents a type with a single value.
+    /// Gets the single <see cref="Void"/> value.
     /// </summary>
-    public readonly struct Void : IEquatable<Void>
-    {
-        /// <summary>
-        /// Gets the single <see cref="Void"/> value.
-        /// </summary>
-        public static Void Default { get; } = default;
+    public static Void Default { get; } = default;
 
-        #region Boilerplate
+    #region Boilerplate
 
-        /// <inheritdoc/>
-        [SuppressMessage("Style", "IDE0060:Remove unused parameter", Justification = "<Pending>")]
-        public static bool operator ==(Void left, Void right) => true;
+    /// <inheritdoc/>
+    [SuppressMessage("Style", "IDE0060:Remove unused parameter", Justification = "<Pending>")]
+    public static bool operator ==(Void left, Void right) => true;
 
-        /// <inheritdoc/>
-        [SuppressMessage("Style", "IDE0060:Remove unused parameter", Justification = "<Pending>")]
-        public static bool operator !=(Void left, Void right) => false;
+    /// <inheritdoc/>
+    [SuppressMessage("Style", "IDE0060:Remove unused parameter", Justification = "<Pending>")]
+    public static bool operator !=(Void left, Void right) => false;
 
-        /// <inheritdoc/>
-        public bool Equals(Void other) => true;
+    /// <inheritdoc/>
+    public bool Equals(Void other) => true;
 
-        /// <inheritdoc/>
-        public override bool Equals([NotNullWhen(true)] object? obj) => obj is Void;
+    /// <inheritdoc/>
+    public override bool Equals([NotNullWhen(true)] object? obj) => obj is Void;
 
-        /// <inheritdoc/>
-        public override int GetHashCode() => 0;
+    /// <inheritdoc/>
+    public override int GetHashCode() => 0;
 
-        /// <inheritdoc/>
-        public override string ToString() => "()";
+    /// <inheritdoc/>
+    public override string ToString() => "()";
 
-        #endregion
-    }
-} 
+    #endregion
+}
