@@ -2,103 +2,101 @@
 using FluentAssertions;
 using System.Text;
 using System.Text.Json;
-using System;
-using System.Collections.Generic;
 
-namespace Pevac.Tests
+namespace Pevac.Tests;
+
+public class ObjectDeserializerTests
 {
-    public class ObjectDeserializerTests
+    private record Foo() { }
+
+    private record Bar() : Foo
     {
-        private record Foo() { }
+        public int BarValue { get; set; }
+    }
 
-        private record Bar() : Foo 
+    private record Baz() : Foo
+    {
+        public string BazValue { get; set; }
+    }
+
+    private record Data(string Foo, string Bar)
+    {
+        public string Foo { get; set; } = Foo;
+        public string Bar { get; set; } = Bar;
+    }
+
+    private record SubType(string Foo, string Bar, string Baz) : Data(Foo, Bar)
+    {
+        public string Baz { get; set; } = Baz;
+
+        public static SubType Copy(Data data) => new(data.Foo, data.Bar, "");
+    }
+
+    private Parser<Foo> FooParser { get; } = Parser.ParseObject(propertyName => propertyName switch
+    {
+        "type" => Parser.String.Updater<string, Foo>(),
+        "bar" => Parser.ParseObject(propertyName => propertyName switch
         {
-            public int BarValue { get; set; }
-        }
-
-        private record Baz(): Foo
+            "bar_value" => Parser.Int32.Updater((int data, Bar bar) => bar.BarValue = data),
+            _ => Parser.FailUpdate<Bar>()
+        }, (Foo parent) => new Bar()),
+        "baz" => Parser.ParseObject(propertyName => propertyName switch
         {
-            public string BazValue { get; set; }
-        }
+            "baz_value" => Parser.String.Updater((string data, Baz baz) => baz.BazValue = data),
+            _ => Parser.FailUpdate<Baz>()
+        }, (Foo _) => new Baz()),
+        _ => Parser.FailUpdate<Foo>()
+    });
 
-        private record Data(string Foo, string Bar)
-        {
-            public string Foo { get; set; } = Foo;
-            public string Bar { get; set; } = Bar;
-        }
-
-        private record SubType(string Foo, string Bar, string Baz): Data(Foo, Bar) 
-        {
-            public string Baz { get; set; } = Baz;
-
-            public static SubType Copy(Data data) => new(data.Foo, data.Bar, "");
-        }
-
-        private Parser<Foo> FooParser { get; } = Parser.ParseObject(propertyName => propertyName switch
-        {
-            "type" => Parser.String.Updater<string, Foo>(),
-            "bar" => Parser.ParseObject(propertyName => propertyName switch
-            {
-                "bar_value" => Parser.Int32.Updater((int data, Bar bar) => bar.BarValue = data),
-                _ => Parser.FailUpdate<Bar>()
-            }, (Foo parent) => new Bar()),
-            "baz" => Parser.ParseObject(propertyName => propertyName switch
-            {
-                "baz_value" => Parser.String.Updater((string data, Baz baz) => baz.BazValue = data),
-                _ => Parser.FailUpdate<Baz>()
-            }, (Foo _) => new Baz()),
-            _ => Parser.FailUpdate<Foo>()
-        });
-
-        [Fact]
-        public void DeserializeObject()
-        {
-            var json = @"
+    [Fact]
+    public void DeserializeObject()
+    {
+        var json = @"
             {
                 ""foo"": ""foo"",
                 ""bar"": ""bar""
             }";
 
-            var bytes = Encoding.UTF8.GetBytes(json);
-            var reader = new Utf8JsonReader(bytes);
-            Parser<Data> parser = Parser.ParseObject(propertyName => propertyName switch
-            {
-                "foo" => Parser.String.Updater((string foo, Data data) => data with { Foo = foo }),
-                "bar" => Parser.String.Updater((string bar, Data data) => data.Bar = bar),
-                _ => Parser.FailUpdate<Data>()
-            }, new Data("", ""));
-
-            Parser.Parse(parser, ref reader, default).Should().Be(new Data("foo", "bar"));
-        }
-
-        [Fact]
-        public void DeserializeStartedObject()
+        var bytes = Encoding.UTF8.GetBytes(json);
+        var reader = new Utf8JsonReader(bytes);
+        Parser<Data> parser = Parser.ParseObject(propertyName => propertyName switch
         {
-            var json = @"
+            "foo" => Parser.String.Updater((string foo, Data data) => data with { Foo = foo }),
+            "bar" => Parser.String.Updater((string bar, Data data) => data.Bar = bar),
+            _ => Parser.FailUpdate<Data>()
+        }, new Data("", ""));
+
+        Parser.Parse(parser, ref reader, default).Should().Be(new Data("foo", "bar"));
+    }
+
+    [Fact]
+    public void DeserializeStartedObject()
+    {
+        var json = @"
             {
                 ""foo"": ""foo"",
                 ""bar"": ""bar""
             }";
 
-            var bytes = Encoding.UTF8.GetBytes(json);
-            var reader = new Utf8JsonReader(bytes);
-            reader.Read();
-            Parser<Data> parser = Parser.ParseObject(propertyName => propertyName switch
-            {
-                "foo" => Parser.String.Updater((string foo, Data data) => data with { Foo = foo }),
-                "bar" => Parser.String.Updater((string bar, Data data) => data.Bar = bar),
-                _ => Parser.FailUpdate<Data>()
-            }, new Data("", ""));
-
-            Parser.Parse(parser, ref reader, default).Should().Be(new Data("foo", "bar"));
-        }
-
-        
-
-        [Fact]
-        public void DeserializeBar()
+        var bytes = Encoding.UTF8.GetBytes(json);
+        var reader = new Utf8JsonReader(bytes);
+        reader.Read();
+        Parser<Data> parser = Parser.ParseObject(propertyName => propertyName switch
         {
-            var json = @"
+            "foo" => Parser.String.Updater((string foo, Data data) => data with { Foo = foo }),
+            "bar" => Parser.String.Updater((string bar, Data data) => data.Bar = bar),
+            _ => Parser.FailUpdate<Data>()
+        }, new Data("", ""));
+
+        Parser.Parse(parser, ref reader, default).Should().Be(new Data("foo", "bar"));
+    }
+
+
+
+    [Fact]
+    public void DeserializeBar()
+    {
+        var json = @"
             {
                 ""type"": ""bar"",
                 ""bar"": {
@@ -106,17 +104,17 @@ namespace Pevac.Tests
                 }
             }";
 
-            var bytes = Encoding.UTF8.GetBytes(json);
-            var reader = new Utf8JsonReader(bytes);
-            reader.Read();
+        var bytes = Encoding.UTF8.GetBytes(json);
+        var reader = new Utf8JsonReader(bytes);
+        reader.Read();
 
-            Parser.Parse(FooParser, ref reader, default).Should().BeEquivalentTo(new Bar { BarValue = 1});
-        }
+        Parser.Parse(FooParser, ref reader, default).Should().BeEquivalentTo(new Bar { BarValue = 1 });
+    }
 
-        [Fact]
-        public void DeserializeBaz()
-        {
-            var json = @"
+    [Fact]
+    public void DeserializeBaz()
+    {
+        var json = @"
             {
                 ""type"": ""bar"",
                 ""baz"": {
@@ -124,10 +122,9 @@ namespace Pevac.Tests
                 }
             }";
 
-            var bytes = Encoding.UTF8.GetBytes(json);
-            var reader = new Utf8JsonReader(bytes);
-            reader.Read();
-            Parser.Parse(FooParser, ref reader, default).Should().BeEquivalentTo(new Baz { BazValue = "1"});
-        }
+        var bytes = Encoding.UTF8.GetBytes(json);
+        var reader = new Utf8JsonReader(bytes);
+        reader.Read();
+        Parser.Parse(FooParser, ref reader, default).Should().BeEquivalentTo(new Baz { BazValue = "1" });
     }
 }
