@@ -1,12 +1,11 @@
-﻿using System;
+﻿namespace Pevac;
 
-namespace Pevac
+using System;
+
+/// <summary>
+/// An exception that will be thrown in unexpected situations(aka. What a fuck?).
+/// </summary>
+public sealed class ParseException : Exception
 {
-    /// <summary>
-    /// An exception that will be thrown in unexpected situations(aka. What a fuck?).
-    /// </summary>
-    public sealed class ParseException : Exception
-    {
-        internal ParseException(string? message = default) : base(message) { }
-    }
+    internal ParseException(string? message = default) : base(message) { }
 }
